Avoid execution of compiled applications during build.
For a currently unknown reason application that link against
the built glibc shared library do not work correctly.

This is not ideal, and the underlying reason for this failure
should at some point be determined, however as the glibc library
is only built to provide a suitable set of symbols, with the correct
version to link against none of the application ever actually
run using this .so. Given this rather than trying to fully debug
the problem, we work around by avoiding running any compiled
applications during the build process.